requirements.txt
blinker==1.4
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-Mail==0.9.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
oauthlib==2.0.2
requests==2.18.3
requests-oauthlib==0.8.0
six==1.10.0
speedtest-cli==1.0.6
tweepy==3.5.0
urllib3==1.22
Werkzeug==0.12.2
